pkg/omf: Refresh after updating the framework
This commit makes `omf update` refresh the session
after updating the framework, setting a env variable
to flag that packages needs to be updated just after
the refresh, which is done in omf plugin init function.

This is needed because if framework code changes
and the functions are relocated the autoloading
mechanism will not find them anymore, causing the
update process to throw a lot of ugly errors.